pyproject.toml
[project]
name = "esjzone-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "colorama>=0.4.6",
    "loguru>=0.7.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "questionary>=2.0.0",
    "requests>=2.31.0",
    "ruamel.yaml>=0.17.0",
]
